Netsocket: Remove deprecated TCPServer
TCPSocket should be used instead.
