+++ features/netsocket/NetworkStack.h
@@ -244,7 +244,6 @@ protected:
     friend class InternetSocket;
     friend class InternetDatagramSocket;
     friend class TCPSocket;
-    friend class TCPServer;
 
     /** Opens a socket
      *
