+++ features/netsocket/nsapi.h
@@ -39,7 +39,6 @@
 #include "netsocket/Socket.h"
 #include "netsocket/UDPSocket.h"
 #include "netsocket/TCPSocket.h"
-#include "netsocket/TCPServer.h"
 #include "netsocket/TLSSocketWrapper.h"
 #include "netsocket/DTLSSocketWrapper.h"
 #include "netsocket/TLSSocket.h"
